Move DiracProcessManager to avoid test failures
